fix(server)!: remove time_warn! from the public API
This is intended to be an internal macro.
